requirements.txt
asgiref==3.7.2
Django==3.2.20
djangorestframework==3.14.0
pip==22.3.1
pytz==2023.3
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.7.1
wheel==0.38.4
